third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEGaussianBlurElement-svgdom-in-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' property of the SVGFEGaussianBlurElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS blurElement.in1.baseVal is "SourceAlpha"
PASS blurElement.in1.baseVal is "SourceGraphic"
PASS successfullyParsed is true

TEST COMPLETE
